+++ apps/web/src/App.tsx
@@ -42,6 +42,7 @@ import HealthPage from './pages/HealthPage';
 import LocationPage from './pages/LocationPage';
 import PhotosPage from './pages/PhotosPage';
 import OnThisDayPage from './pages/OnThisDayPage';
+import PersonaPlexPage from './pages/PersonaPlexPage';
 import DataReclamationMarketingPage from './pages/DataReclamationMarketingPage';
 import HelpPage from './pages/HelpPage';
 import FamilyPage from './pages/FamilyPage';
@@ -146,6 +147,7 @@ export default function App() {
             <Route path="location" element={<LocationPage />} />
             <Route path="photos" element={<PhotosPage />} />
             <Route path="on-this-day" element={<OnThisDayPage />} />
+            <Route path="persona-plex" element={<PersonaPlexPage />} />
             <Route path="storage-settings" element={<StorageSettingsPage />} />
             <Route path="settings" element={<SettingsPage />} />
             <Route path="family" element={<FamilyPage />} />